Fix compiler warning (#122)
gcc warnings fixed

+++ src/mplatform.cpp
@@ -799,7 +799,7 @@ static inline mst::flag<mst::platform::processor_feature_flags> processor_featur
 
 uint64_t mst::_Details::processor_features_impl() noexcept
 {
-	static uint32_t features = processor_features_init().get();
+	static uint64_t features = processor_features_init().get();
 
 	return features;
 }
